cells: move command support into commons-cli and create skeleton CLI client
This patch moves the classes needed to support a CLI from cells into
commons-cli.

The Interpretable interface is removed as it had no effect.

The CommandRequestable interface is removed.  Its purpose was to allow
arbitrary arguments; however, it was used for a single command (which
was never used in dCache) and the same effect is achievable via
escaping.

The chimera shell is split into two parts: the generic CLI handling part
is now the ShellApplication class in commons-cli, from which the
Chimera shell class now inherits.

Target: master
Requires-notes: no
Requires-book: no
Patch: http://rb.dcache.org/r/6552
Acked-by: Gerd Behrmann