srm-client: fix adler32
The adler32 script invokes the Adler32 class to calculate the ADLER32 digest of a file.

Subsequent to the 2.6.0 release, this class was mistakenly moved to srm-server, so isn't
part of the srm-client packages, which means that this command fails if srm-client is
built from branches newer than 2.6.

This patch moves the Adler32 class to the srm-client package.  It also moves the GridftpList
class that is also used by a script (which we currently don't distribute).  The two scripts
command-line-clietn and srm_v1 (both of which we don't distribute) are deleted as they
refer to classes that do not exist.  The unused Makefile is also deleted.

Target: master
Request: 2.8
Request: 2.7
Acked-by: Gerd Behrmann
Patch: http://rb.dcache.org/r/6643/