gss: tidy up FQAN-extraction
The SRM interface and the dcap door extract a list of FQANs from a
certificate chain.  In the past, the caller was responsible for
supplying the objects used to achieve this; but, some time ago, this
changed and these objects are created within GSSUtils, independent of
the caller.

The interface was not updated to reflect this, so callers were forced
to create objects that were then not used.  Also, methods marked
themselves as throwing an Exception that they never do.

These wrinkles are fixed with this patch.

Target: master
Patch: http://rb.dcache.org/r/6656/
Acked-by: Gerd Behrmann