ftp: Make root path behaviour configurable
Our FTP door is different from all other doors in dCache by
exposing a user-specific root path. This patch introduces
the ftp.root property. It mimics that of webdav.root and
xrootd.root by defining which directory should form the root
of the door. If set to the empty string (which is the default
value), the existing behaviour is maintained.

As before, the user's home directory is used to define the
initial working directory of a new FTP session.

Also fixes a bug that could prevent the user's settings from
being applied if logging in multiple times in the same FTP
session.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6686/