billing: Initialize all fields in billing records
Plenty of billing records have fields that are logged as 'unknown' due to
missing values. This patch injects those missing values.

I added file path to PoolIoFileMessage to allow the pool to inject the path
into the billing record. Although the paht is known as part of StorageInfo,
StorageInfo is the wrong place to store the path: The path is not an attribute
of a file. A file can have many paths pointing to it.  A path is rather an
attribute of a particular access (the file is accessed through a path) and thus
the path should be part of the request messages. For the same reason a path
should not become an attribute in FileAttributes.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6689/