ftp: add support for HELP and SITE HELP commands
Both the ncftp and Globus clients parse the output of HELP commands
to determine which additional commands are available.  Both use
broadly the same approach: each response line is searched for
references to any of the SITE commands the client understands.  The
search is a simple substring search, looking for the command in
upper case, excluding the initial 'SITE '.  The ncftp client parses
the output of 'HELP SITE' and the Globus client parses the output
of 'SITE HELP'.

This patch is to allow support for additional commands.

Target: master
Patch: http://rb.dcache.org/r/6701
Acked-by: Gerd Behrmann