ftp: Fix race condition in command dispatching
Fixes a race condition that lead to hanging FTP sessions.

A FutureTask is not thread safe in the sense that it cannot be
submitted concurrently for several executions.

I must admit, I have no idea why the code was like it was. The future
must have been part of an earlier revision of the code.

Target: trunk
Request: 2.8
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6704/