ftp: fix printing of UNIX.Mode fact
The format of the UNIX.Mode fact is not documented.  The ncftp client
expacts that a value starting '0' is in octal and other values are in
decimal.  The Globus server follows this convention by returning an
octal number that starts '0'.

Currently dCache returns an octal number without the leading '0'.
This patch fixes this by appending the '0'.

Target: master
Patch: http://rb.dcache.org/r/6639/
Acked-by: Tigran Mkrtchyan