Add transparent retry support to async use of CellStub
Synchronous message passing has support an option to transparently
resend messages on NoRouteToCellException. This is to be used for
destinations that are known to be required in a dCache installation
(like PnfsManager) when no sensible alternative action exists.

Support has been limited to the synchronous sendAndWait methods. This
patch adds support for the callback based sendMessage calls. On failure
to deliver a message, the message is not resend right away to avoid
a tight retry loop. Instead the message is send while performing
other cell background tasks like processing cell message timeouts.

The change is made in anticipation of a future patch that refactors
the sendAndWait calls to be build upon the async infrastructure of
CellStub.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6739/