ftp: add support for updating atime
This patch adds support for a client to update a file's or directory's
atime.  The MFMT, MFCT and MFF commands requires that the actual value
is returned to the client, allowing that this might not be the value
the client requested.

When updating atime, the actual value is somewhat unpredicable as it
depends on the pnfsmanager.atime-gap value and the previous atime
value.  Therefore, to return the correct value, the FTP door must query
chimera to find the actual value set.

To avoid the overhead of an extra RTT, this patch modifies the behavour
of the PnfsSetFileAttributes so that it may optionally return a set of
attributes.  The default behaviour queries none, which results in no
additional overhead.

The FTP door is updated so MFMT, MFCT and MFF commands return the
actual value set rather than assuming the user-supplied value is
accepted.  This allows these commands to return more robust response
values to the client.

Target: master
Patch: http://rb.dcache.org/r/6802/
Acked-by: Gerd Behrmann
Acked-by: Tigran Mkrtchyan