loginbroker: Fix login broker registration retry
Fixes a problem in which doors would log "Failed to send update to LoginBroker"
in a busy retry loop when unable to register with LoginBroker. This typically
happens during shutdown.

The intention of the code was to switch to an eager retry loop with a one
second sleep between retries, however the code switches the update period to
'normal' update before attempting to send an update and then switches back to
'eager' mode. Changing the scheduling time causes the task to be rescheduled,
but rescheduling schedules a new update with an initial delay of zero, thus
causing an immediate retry. Thus failure to deliver an update initiates a
tight retry loop without any delays.

Also fixes a bug in which the 'lb set update' command was not respected.

Target: trunk
Request: 2.8
Request: 2.7
Request: 2.6
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6787/