logback: add comment about bugs in 1.1.1 to poms
We are not quite ready to update to 1.1.1 yet because of the
issue which persists with respect to main not exiting in the
alarm send command.

I have added comments to the poms indicating the bug reports
I filed with logback.

Target: master
Patch: http://rb.dcache.org/r/6786/
Require-notes: no
Require-book: no
Acked-by: Gerd