spacemanager: Fix DB trigger migration
The space manager liquibase files follow liquibase best practice
recommendations by placing stored procedures in a separate changeset.
Unfortunately, for triggers this is a bad advice: When making schema changes,
it is not uncommon to run update or delete queries that would cause the
triggers to execute. Thus trigger updates have to be carefully interleaved with
other schema changes.

This patch also resolves a typo in the 2.9 schema changes affecting the ability
to rollback the changes.

Let me stress that the liquibase best practice is still good advice: It should
however only be applied to stored procedures used by the application, not to
triggers.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6814/