srm: fix restore from database
Many (but not all) ContainerRequests are not scheduled and rely on
the scheduling of the member FileRequest jobs.  These receive a
scheduler for their FileRequest jobs, which they cannot use should
they wish to schedule themselves.

The current restore-after-restart code mistakenly tries to reschedule
these ContainerRequest jobs with their member FileRequest scheduler.
This fails as the container jobs are the wrong type for the scheduler.

This patch fixes this issue by suppressing the default behaviour of
scheduling jobs for container jobs.

Target: master
Patch: http://rb.dcache.org/r/6815
Acked-by: Gerd Behrmann