srm: Avoid loading the same job multiple times
Restore from the database is suppossed to be fairly ordered: First the file
requests are read and added to an in-memory store and then the container
requests are read and an aggregation of the file requests is formed.

Unfortunately while reading the file requests, a debug log statement calls
Job#getSubmitterId which triggers an unexpected load of the container request.
This severily messes up restore from the database.

The patch removes this log statement. It also reduces some of the overhead of
loading the job history.

Target: trunk
Request: 2.8
Request: 2.7
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6820/