srm: fix restore so jobs for other schedulers are not restored
The dCache may have multiple SRM instances and jobs are associated
with a specific scheduler.

Currently on startup, jobs may be restored that are associated with
other SRM instances.  This may lead to the job being processed
multiple times, leading to inconsistent results.

This patch alters the behaviour so only jobs for the named scheduler
will be started.  It also adds a check that jobs that have been
assigned a scheduler are submitted only to that scheduler.

Target: master
Patch: http://rb.dcache.org/r/6826/
Acked-by: Gerd Behrmann