cells: Fix message timeout
CellNucleus relies on periodic calls to executeMaintenanceTasks for message
timeouts for asynchronous tasks to work. Every dCache cell does this
automatically by inheriting from AbstractCell, but old-scholl cells in the
cells package don't. This was not a problem as long as cells only used blocking
communication, however a recent change refactored blocking communication on top
of asynchronous communication. This broke reconnect in LocationManager.

The patch addresses the problem on two fronts: It adds an explicit timeout to
the blocking get call in sendAndWait, thus ensuring a timely timeout; and it
moves the automatic calls of executeMintenanceTasks from AbstractCell to
CellNucleus.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6843/