gplazma-nis: suppress annoying error message
If the NIS plugin fails to map a username then it logs this as an error.

On gPlazma configured kerberos and NIS, where the Kerberos host
principal is used when mounting via NFS, there are many times where
NIS plugin is unable to map the host credential.  This results in
the plugin spamming the log file with meaningless error messages.

This patch pushes the significance of this to debug level.

Target: master
Patch: http://rb.dcache.org/r/6845/
Acked-by: Gerd Behrmann