ftp: add work-around for GlobusOnline uploads
The GlobusOnline client expects that the MD5 checksum is available
after uploading the file and does not degrade gracefully if an
alternative checksum is supplied.

This patch adds support where, if the GO client identifies itself,
then the pool is requested to generate an MD5 checksum.

Target: master
Patch: http://rb.dcache.org/r/6853
Acked-by: Gerd Behrmann