srm: make timestamps easier to read
The current format for time-stamps in SRM output follows ISO 8601, but is
needlessly verbose.

The timezone designator is optional: if ommitted, local timezone is assumed.
This only matters if the information is shared across multiple timezones and
the source is ambigous.  Since log files also include the SRM's cell name,
which is unique within a dCache inistance, I believe such ambiguity is
sufficiently unlikely that it doesn't justify the extra three characters in the
output.

According to Wikipedia, ISO 8601 says:

    4.3.2 NOTE: By mutual agreement of the partners in information interchange,
          the character [T] may be omitted in applications where there is no
          risk of confusing a date and time of day representation with others
          defined in this International Standard.

Since (personally) I find the 'T' makes the output harder to read and a single
space easier, this patch removes the 'T' and replaces it with a single space.

I'm not sure if this is in complete agreement with the above note, but I
believe there's no real risk of misinterpretation the fields.

Target: master